+++ packages/framer-motion/src/render/utils/KeyframesResolver.ts
@@ -69,6 +69,7 @@ function measureAllKeyframes() {
 }
 
 function readAllKeyframes() {
+    console.log("reading all keyframes", toResolve.size)
     toResolve.forEach((resolver) => {
         resolver.readKeyframes()